refactor(sqlsmith): Separate generation of functions, agg, cast from expr (#10008)
