+++ src/tests/sqlsmith/src/sql_gen/mod.rs
@@ -23,10 +23,13 @@ use risingwave_common::types::DataType;
 use risingwave_frontend::bind_data_type;
 use risingwave_sqlparser::ast::{ColumnDef, Expr, Ident, ObjectName, Statement};
 
+mod agg;
+mod cast;
 mod expr;
 pub use expr::print_function_table;
 
 mod dml;
+mod functions;
 mod query;
 mod relation;
 mod scalar;
